controller manager

+++ Assets/Scripts/MachankoHandler.cs
@@ -1,10 +1,12 @@
+using UnityEngine;
+
 public class MachankoHandler : MonoBehaviour
 {
     void Awake(){
         HandTracker.OnMachanko += Bless;
     }
 
-    void Bless(int force){
+    void Bless(float force){
         
     }
 }
